Extract common matview model methods to a concern
