+++ app/models/concerns/patient_reportable.rb
@@ -1,6 +1,6 @@
 module PatientReportable
   extend ActiveSupport::Concern
-  LTFU_PERIOD = 12.months
+  TIME_TO_CONSIDER_LTFU = 12.months
 
   included do
     scope :with_diabetes, -> { joins(:medical_history).merge(MedicalHistory.diabetes_yes).distinct }
@@ -9,12 +9,12 @@ module PatientReportable
 
     scope :ltfu_as_of, ->(date) do
       where.not(id: latest_bps_within_ltfu_period(date).select(:patient_id).distinct)
-        .where("recorded_at < ?", date - LTFU_PERIOD)
+        .where("recorded_at < ?", date - TIME_TO_CONSIDER_LTFU)
     end
 
     scope :not_ltfu_as_of, ->(date) do
       where(id: latest_bps_within_ltfu_period(date).select(:patient_id).distinct)
-        .or(where("patients.recorded_at > ?", date - LTFU_PERIOD))
+        .or(where("patients.recorded_at > ?", date - TIME_TO_CONSIDER_LTFU))
     end
 
     scope :for_reports, ->(with_exclusions: false, exclude_ltfu_as_of: nil) do
@@ -34,7 +34,7 @@ module PatientReportable
 
     def self.latest_bps_within_ltfu_period(ltfu_as_of)
       LatestBloodPressuresPerPatientPerMonth
-        .where("bp_recorded_at > ? AND bp_recorded_at <= ?", ltfu_as_of - LTFU_PERIOD, ltfu_as_of)
+        .where("bp_recorded_at > ? AND bp_recorded_at <= ?", ltfu_as_of - TIME_TO_CONSIDER_LTFU, ltfu_as_of)
     end
   end
 end
