+++ app/models/concerns/patient_reportable_matview.rb
@@ -1,0 +1,24 @@
+module PatientReportableMatview
+  extend ActiveSupport::Concern
+  extend PatientReportable
+  TIME_TO_CONSIDER_LTFU = PatientReportable::TIME_TO_CONSIDER_LTFU
+
+  included do
+    def self.refresh
+      Scenic.database.refresh_materialized_view(table_name, concurrently: true, cascade: false)
+    end
+
+    scope :with_hypertension, -> { where("medical_history_hypertension = ?", "yes") }
+    scope :excluding_dead, -> { where.not(patient_status: :dead) }
+
+    scope :ltfu_as_of, ->(date) do
+      where.not("bp_recorded_at > ? AND bp_recorded_at <= ?", date.to_date - TIME_TO_CONSIDER_LTFU, date.to_date)
+           .where("patient_recorded_at < ?", date - TIME_TO_CONSIDER_LTFU)
+    end
+
+    scope :not_ltfu_as_of, ->(date) do
+      where("bp_recorded_at > ? AND bp_recorded_at <= ?", date.to_date - TIME_TO_CONSIDER_LTFU, date.to_date)
+        .or(where("patient_recorded_at > ?", date - TIME_TO_CONSIDER_LTFU))
+    end
+  end
+end
