+++ app/models/latest_blood_pressures_per_patient.rb
@@ -1,12 +1,8 @@
 class LatestBloodPressuresPerPatient < ApplicationRecord
   include BloodPressureable
-  include PatientReportable
+  include PatientReportableMatview
 
-  def self.refresh
-    Scenic.database.refresh_materialized_view(table_name, concurrently: true, cascade: false)
-  end
-
+  belongs_to :patient
   belongs_to :bp_facility, class_name: "Facility", foreign_key: :bp_facility_id
   belongs_to :registration_facility, class_name: "Facility", foreign_key: :registration_facility_id
-  belongs_to :patient
 end
